# Speedup FP4 packing