@wenhuach21 2025-08-25
From our tests with the AutoRound format, torch compile only provides noticeable benefits for large models or MoE models, since compilation takes additional time. For a 7B model, packing with torch compile typically takes only about 15 seconds, but with torch compile, it costs 30s